+++ lib/src/configuration/better_player_data_source.dart
@@ -1,6 +1,6 @@
 // Project imports:
 
-import 'package:better_player/src/configuration/better_player_android_configuration.dart';
+import 'package:better_player/src/configuration/better_player_buffering_configuration.dart';
 import 'package:better_player/src/configuration/better_player_data_source_type.dart';
 import 'package:better_player/src/configuration/better_player_drm_configuration.dart';
 import 'package:better_player/src/configuration/better_player_notification_configuration.dart';
@@ -74,8 +74,9 @@ class BetterPlayerDataSource {
   /// BetterPlayerConfiguration.
   final Widget? placeholder;
 
-  ///Specific android related code which enables the user to f.e. optimize video loading, buffering and playing
-  final BetterPlayerAndroidConfiguration betterPlayerAndroidConfiguration;
+  ///Configuration of video buffering. Currently only supported in Android
+  ///platform.
+  final BetterPlayerBufferingConfiguration bufferingConfiguration;
 
   BetterPlayerDataSource(
     this.type,
@@ -90,16 +91,21 @@ class BetterPlayerDataSource {
     this.asmsTrackNames,
     this.resolutions,
     this.cacheConfiguration,
-    this.betterPlayerAndroidConfiguration = const BetterPlayerAndroidConfiguration(),
-    this.notificationConfiguration = const BetterPlayerNotificationConfiguration(showNotification: false),
+    this.notificationConfiguration =
+        const BetterPlayerNotificationConfiguration(
+      showNotification: false,
+    ),
     this.overriddenDuration,
     this.videoFormat,
     this.videoExtension,
     this.drmConfiguration,
     this.placeholder,
+    this.bufferingConfiguration = const BetterPlayerBufferingConfiguration(),
   }) : assert(
-            (type == BetterPlayerDataSourceType.network || type == BetterPlayerDataSourceType.file) ||
-                (type == BetterPlayerDataSourceType.memory && bytes?.isNotEmpty == true),
+            (type == BetterPlayerDataSourceType.network ||
+                    type == BetterPlayerDataSourceType.file) ||
+                (type == BetterPlayerDataSourceType.memory &&
+                    bytes?.isNotEmpty == true),
             "Url can't be null in network or file data source | bytes can't be null when using memory data source");
 
   ///Factory method to build network data source which uses url as data source
@@ -114,13 +120,14 @@ class BetterPlayerDataSource {
     bool? useAsmsAudioTracks,
     Map<String, String>? qualities,
     BetterPlayerCacheConfiguration? cacheConfiguration,
-    BetterPlayerAndroidConfiguration betterPlayerAndroidConfiguration = const BetterPlayerAndroidConfiguration(),
     BetterPlayerNotificationConfiguration notificationConfiguration =
         const BetterPlayerNotificationConfiguration(showNotification: false),
     Duration? overriddenDuration,
     BetterPlayerVideoFormat? videoFormat,
     BetterPlayerDrmConfiguration? drmConfiguration,
     Widget? placeholder,
+    BetterPlayerBufferingConfiguration bufferingConfiguration =
+        const BetterPlayerBufferingConfiguration(),
   }) {
     return BetterPlayerDataSource(
       BetterPlayerDataSourceType.network,
@@ -138,7 +145,7 @@ class BetterPlayerDataSource {
       videoFormat: videoFormat,
       drmConfiguration: drmConfiguration,
       placeholder: placeholder,
-      betterPlayerAndroidConfiguration: betterPlayerAndroidConfiguration
+      bufferingConfiguration: bufferingConfiguration,
     );
   }
 
@@ -220,6 +227,8 @@ class BetterPlayerDataSource {
     String? videoExtension,
     BetterPlayerDrmConfiguration? drmConfiguration,
     Widget? placeholder,
+    BetterPlayerBufferingConfiguration? bufferingConfiguration =
+        const BetterPlayerBufferingConfiguration(),
   }) {
     return BetterPlayerDataSource(
       type ?? this.type,
@@ -233,12 +242,15 @@ class BetterPlayerDataSource {
       useAsmsAudioTracks: useAsmsAudioTracks ?? this.useAsmsAudioTracks,
       resolutions: resolutions ?? this.resolutions,
       cacheConfiguration: cacheConfiguration ?? this.cacheConfiguration,
-      notificationConfiguration: notificationConfiguration ?? this.notificationConfiguration,
+      notificationConfiguration:
+          notificationConfiguration ?? this.notificationConfiguration,
       overriddenDuration: overriddenDuration ?? this.overriddenDuration,
       videoFormat: videoFormat ?? this.videoFormat,
       videoExtension: videoExtension ?? this.videoExtension,
       drmConfiguration: drmConfiguration ?? this.drmConfiguration,
       placeholder: placeholder ?? this.placeholder,
+      bufferingConfiguration:
+          bufferingConfiguration ?? this.bufferingConfiguration,
     );
   }
 }