+++ lib/src/core/better_player_controller.dart
@@ -50,10 +50,6 @@ class BetterPlayerController {
   ///List of event listeners, which listen to events.
   final List<Function(BetterPlayerEvent)?> _eventListeners = [];
 
-  ///User is able to check wether there are already listeners registered and if so can remove them
-  ///Is useful in cases where multiple listeners should not be allowed and therefore need to be removed
-  List<Function(BetterPlayerEvent)?> get eventListeners => _eventListeners;
-
   ///List of files to delete once player disposes.
   final List<File> _tempFiles = [];
 
@@ -246,7 +242,9 @@ class BetterPlayerController {
 
     ///Build videoPlayerController if null
     if (videoPlayerController == null) {
-      videoPlayerController = VideoPlayerController(betterPlayerAndroidConfiguration: betterPlayerDataSource.betterPlayerAndroidConfiguration);
+      videoPlayerController = VideoPlayerController(
+          bufferingConfiguration:
+              betterPlayerDataSource.bufferingConfiguration);
       videoPlayerController?.addListener(_onVideoPlayerChanged);
     }
 
@@ -462,7 +460,8 @@ class BetterPlayerController {
           formatHint: _getVideoFormat(_betterPlayerDataSource!.videoFormat),
           licenseUrl: _betterPlayerDataSource?.drmConfiguration?.licenseUrl,
           certificateUrl:
-                _betterPlayerDataSource?.drmConfiguration?.certificateUrl,drmHeaders: _betterPlayerDataSource?.drmConfiguration?.headers,
+              _betterPlayerDataSource?.drmConfiguration?.certificateUrl,
+          drmHeaders: _betterPlayerDataSource?.drmConfiguration?.headers,
           activityName:
               _betterPlayerDataSource?.notificationConfiguration?.activityName,
         );