+++ lib/src/subtitles/better_player_subtitles_drawer.dart
@@ -92,10 +92,10 @@ class _BetterPlayerSubtitlesDrawerState
   ///Called when player state has changed, i.e. new player position, etc.
   void _updateState() {
     if (mounted) {
-      /*setState(() {*/
+      setState(() {
         _latestValue =
             widget.betterPlayerController.videoPlayerController!.value;
-      /*});*/
+      });
     }
   }